+++ ccd-definitions/jurisdictions/england-wales/json/CaseEvent/CaseEvent.json
@@ -229,7 +229,7 @@
     "DisplayOrder": 53,
     "PreConditionState(s)": "Accepted",
     "PostConditionState": "*",
-    "CallBackURLAboutToSubmitEvent": "${ET_COS_URL}/et3Response/aboutToSubmitAmendRepresentativeContact",
+    "CallBackURLAboutToSubmitEvent": "${ET_COS_URL}/amendRepresentativeContact/aboutToSubmit",
     "SecurityClassification": "Public",
     "ShowEventNotes": "N",
     "ShowSummary": "Y"
