+++ ccd-definitions/jurisdictions/england-wales/json/CaseEventToFields/CaseEventToFields.json
@@ -4892,7 +4892,7 @@
     "PageID": 1,
     "PageDisplayOrder": 1,
     "PageFieldDisplayOrder": 1,
-    "CallBackURLMidEvent": "${ET_COS_URL}/et3Response/midEventAmendRepresentativeContact",
+    "CallBackURLMidEvent": "${ET_COS_URL}/amendRepresentativeContact/midEvent",
     "ShowSummaryChangeOption": "N"
   },
   {
