+++ ccd-definitions/jurisdictions/scotland/json/CaseEvent/CaseEvent.json
@@ -171,7 +171,7 @@
     "SecurityClassification": "Public",
     "ShowEventNotes": "N",
     "ShowSummary": "Y",
-    "CallBackURLAboutToSubmitEvent": "${ET_COS_URL}/et3Response/aboutToSubmitAmendRepresentativeContact"
+    "CallBackURLAboutToSubmitEvent": "${ET_COS_URL}/amendRepresentativeContact/aboutToSubmit"
   },
   {
     "CaseTypeID": "ET_Scotland",
