+++ ccd-definitions/jurisdictions/scotland/json/CaseEventToFields/CaseEventToFields.json
@@ -5074,7 +5074,7 @@
     "PageID": 1,
     "PageDisplayOrder": 1,
     "PageFieldDisplayOrder": 1,
-    "CallBackURLMidEvent": "${ET_COS_URL}/et3Response/midEventAmendRepresentativeContact",
+    "CallBackURLMidEvent": "${ET_COS_URL}/amendRepresentativeContact/midEvent",
     "ShowSummaryChangeOption": "N"
   },
   {
