+++ src/test/java/uk/gov/hmcts/ethos/replacement/docmosis/controllers/Et3ResponseControllerTest.java
@@ -1,13 +1,10 @@
 package uk.gov.hmcts.ethos.replacement.docmosis.controllers;
 
-import lombok.SneakyThrows;
-import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
@@ -25,6 +22,7 @@ import uk.gov.hmcts.ethos.replacement.docmosis.exceptions.GenericServiceExceptio
 import uk.gov.hmcts.ethos.replacement.docmosis.service.CaseManagementForCaseWorkerService;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.Et3ResponseService;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.FeatureToggleService;
+import uk.gov.hmcts.ethos.replacement.docmosis.service.noc.AmendRepresentativeContactService;
 import uk.gov.hmcts.ethos.replacement.docmosis.utils.JsonMapper;
 import uk.gov.hmcts.ethos.utils.CCDRequestBuilder;
 import uk.gov.hmcts.ethos.utils.CaseDataBuilder;
@@ -42,14 +40,12 @@ import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.ENGLANDWALES_CASE_TYPE_ID;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.NO;
 import static uk.gov.hmcts.ethos.replacement.docmosis.constants.ET3ResponseConstants.ERROR_CASE_DATA_NOT_FOUND;
-import static uk.gov.hmcts.ethos.replacement.docmosis.constants.ET3ResponseConstants.REPRESENTATIVE_CONTACT_CHANGE_OPTION_USE_MYHMCTS_DETAILS;
 import static uk.gov.hmcts.ethos.replacement.docmosis.helpers.Et3ResponseHelper.ET3_RESPONSE;
 
 @ExtendWith(SpringExtension.class)
@@ -69,15 +65,13 @@ class Et3ResponseControllerTest extends BaseControllerTest {
     private static final String DOWNLOAD_DRAFT_ABOUT_TO_START = "/et3Response/downloadDraft/aboutToStart";
     private static final String DOWNLOAD_DRAFT_ABOUT_TO_SUBMIT = "/et3Response/downloadDraft/aboutToSubmit";
     private static final String DOWNLOAD_DRAFT_SUBMITTED = "/et3Response/downloadDraft/submitted";
-    private static final String ABOUT_TO_SUBMIT_AMEND_REPRESENTATIVE_CONTACT =
-            "/et3Response/aboutToSubmitAmendRepresentativeContact";
-    private static final String MID_EVENT_AMEND_REPRESENTATIVE_CONTACT =
-            "/et3Response/midEventAmendRepresentativeContact";
 
     @Autowired
     private WebApplicationContext applicationContext;
     @MockitoBean
     private Et3ResponseService  et3ResponseService;
+    @MockitoBean
+    private AmendRepresentativeContactService amendRepresentativeContactService;
     @MockitoBean
     private CaseManagementForCaseWorkerService caseManagementForCaseWorkerService;
     private MockMvc mvc;
@@ -311,7 +305,7 @@ class Et3ResponseControllerTest extends BaseControllerTest {
     void submitSection_withEt3ResponseEvent_setsRepresentativeContactDetails() throws Exception {
         when(verifyTokenService.verifyTokenSignature(AUTH_TOKEN)).thenReturn(true);
         ccdRequest.setEventId(ET3_RESPONSE);
-        doNothing().when(et3ResponseService).setRespondentRepresentsContactDetails(
+        doNothing().when(amendRepresentativeContactService).setRespondentRepresentsContactDetails(
                 anyString(), any(CaseData.class), anyString());
 
         mvc.perform(post(SUBMIT_SECTION_URL)
@@ -322,7 +316,7 @@ class Et3ResponseControllerTest extends BaseControllerTest {
                 .andExpect(jsonPath(JsonMapper.DATA, notNullValue()))
                 .andExpect(jsonPath(JsonMapper.WARNINGS, nullValue()));
 
-        verify(et3ResponseService).setRespondentRepresentsContactDetails(
+        verify(amendRepresentativeContactService).setRespondentRepresentsContactDetails(
                 anyString(), any(CaseData.class), anyString());
     }
 
@@ -336,7 +330,7 @@ class Et3ResponseControllerTest extends BaseControllerTest {
                 ccdRequest.getCaseDetails().getCaseId(),
                 "Et3ResponseService",
                 "setRespondentRepresentsContactDetails"))
-                .when(et3ResponseService).setRespondentRepresentsContactDetails(
+                .when(amendRepresentativeContactService).setRespondentRepresentsContactDetails(
                         anyString(), any(CaseData.class), anyString());
 
         mvc.perform(post(SUBMIT_SECTION_URL)
@@ -501,88 +495,6 @@ class Et3ResponseControllerTest extends BaseControllerTest {
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath(JsonMapper.DATA, notNullValue()))
-                .andExpect(jsonPath(JsonMapper.WARNINGS, nullValue()));
-    }
-
-    @Test
-    @SneakyThrows
-    void theMidEventAmendRepresentativeContact() {
-        when(verifyTokenService.verifyTokenSignature(AUTH_TOKEN)).thenReturn(true);
-        doNothing().when(et3ResponseService).setRespondentRepresentsContactDetails(AUTH_TOKEN,
-                ccdRequest.getCaseDetails().getCaseData(), ccdRequest.getCaseDetails().getCaseId());
-        ccdRequest.getCaseDetails().getCaseData().setRepresentativeContactChangeOption(
-                REPRESENTATIVE_CONTACT_CHANGE_OPTION_USE_MYHMCTS_DETAILS);
-        mvc.perform(post(MID_EVENT_AMEND_REPRESENTATIVE_CONTACT)
-                        .contentType(APPLICATION_JSON)
-                        .header(HttpHeaders.AUTHORIZATION, AUTH_TOKEN)
-                        .content(jsonMapper.toJson(ccdRequest)))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath(JsonMapper.DATA, notNullValue()))
-                .andExpect(jsonPath("$.errors.size()", is(0)))
-                .andExpect(jsonPath(JsonMapper.WARNINGS, nullValue()));
-        ccdRequest.getCaseDetails().getCaseData().setRepresentativeContactChangeOption("DUMMY");
-    }
-
-    @Test
-    @SneakyThrows
-    void theMidEventAmendRepresentativeContact_WithException() {
-        when(verifyTokenService.verifyTokenSignature(AUTH_TOKEN)).thenReturn(true);
-        doThrow(new GenericServiceException(ERROR_CASE_DATA_NOT_FOUND,
-                new Exception(ERROR_CASE_DATA_NOT_FOUND),
-                ERROR_CASE_DATA_NOT_FOUND,
-                StringUtils.EMPTY,
-                "Et3ResponseService",
-                "setRespondentRepresentsContactDetails")).when(et3ResponseService)
-                .setRepresentativeMyHmctsContactAddress(anyString(), any(CaseData.class), anyString());
-        ccdRequest.getCaseDetails().getCaseData().setRepresentativeContactChangeOption(
-                REPRESENTATIVE_CONTACT_CHANGE_OPTION_USE_MYHMCTS_DETAILS);
-        mvc.perform(post(MID_EVENT_AMEND_REPRESENTATIVE_CONTACT)
-                        .contentType(APPLICATION_JSON)
-                        .header(HttpHeaders.AUTHORIZATION, AUTH_TOKEN)
-                        .content(jsonMapper.toJson(ccdRequest)))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath(JsonMapper.DATA, notNullValue()))
-                .andExpect(jsonPath("$.errors.size()", is(1)))
-                .andExpect(jsonPath("$.errors[0]", is(ERROR_CASE_DATA_NOT_FOUND)))
-                .andExpect(jsonPath(JsonMapper.WARNINGS, nullValue()));
-        ccdRequest.getCaseDetails().getCaseData().setRepresentativeContactChangeOption("DUMMY");
-    }
-
-    @Test
-    @SneakyThrows
-    void theAboutToSubmitRepresentativeContactDetails() {
-        when(verifyTokenService.verifyTokenSignature(AUTH_TOKEN)).thenReturn(true);
-        doNothing().when(et3ResponseService).setRespondentRepresentsContactDetails(AUTH_TOKEN,
-                ccdRequest.getCaseDetails().getCaseData(), ccdRequest.getCaseDetails().getCaseId());
-        mvc.perform(post(ABOUT_TO_SUBMIT_AMEND_REPRESENTATIVE_CONTACT)
-                        .contentType(APPLICATION_JSON)
-                        .header(HttpHeaders.AUTHORIZATION, AUTH_TOKEN)
-                        .content(jsonMapper.toJson(ccdRequest)))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath(JsonMapper.DATA, notNullValue()))
-                .andExpect(jsonPath("$.errors.size()", is(0)))
-                .andExpect(jsonPath(JsonMapper.WARNINGS, nullValue()));
-    }
-
-    @Test
-    @SneakyThrows
-    void theAboutToSubmitRepresentativeContactDetails_WithException() {
-        when(verifyTokenService.verifyTokenSignature(AUTH_TOKEN)).thenReturn(true);
-        doThrow(new GenericServiceException(ERROR_CASE_DATA_NOT_FOUND,
-                new Exception(ERROR_CASE_DATA_NOT_FOUND),
-                ERROR_CASE_DATA_NOT_FOUND,
-                StringUtils.EMPTY,
-                "Et3ResponseService",
-                "setRespondentRepresentsContactDetails")).when(et3ResponseService)
-                .setRespondentRepresentsContactDetails(anyString(), any(CaseData.class), anyString());
-        mvc.perform(post(ABOUT_TO_SUBMIT_AMEND_REPRESENTATIVE_CONTACT)
-                        .contentType(APPLICATION_JSON)
-                        .header(HttpHeaders.AUTHORIZATION, AUTH_TOKEN)
-                        .content(jsonMapper.toJson(ccdRequest)))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath(JsonMapper.DATA, notNullValue()))
-                .andExpect(jsonPath("$.errors.size()", is(1)))
-                .andExpect(jsonPath("$.errors[0]", is(ERROR_CASE_DATA_NOT_FOUND)))
                 .andExpect(jsonPath(JsonMapper.WARNINGS, nullValue()));
     }
 }
